requirements.txt
flask-cors==5.0.1
Flask==3.1.0
flask_cors==5.0.1
gunicorn==23.0.0
openai==1.70.0
pandas==2.2.3
pydub==0.25.1
python-dotenv==1.1.0
Requests==2.32.3
scipy==1.15.2
tiktoken==0.9.0
